rename `parent-gametes.js` to `gametes.js`
